style changes, move out golay.c
