+++ lib/f_core/radio/protocols/horus/CMakeLists.txt
@@ -2,5 +2,5 @@
 # SPDX-License-Identifier: Apache-2.0
 
 zephyr_library()
-zephyr_library_sources(horus.c)
+zephyr_library_sources(horus.c golay.c)
