+++ lib/f_core/radio/protocols/horus/horus.c
@@ -1,10 +1,13 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
+#include "f_core/radio/protocols/horus/horus.h"
 
+#include "f_core/radio/protocols/horus/golay.h"
+
+#include <assert.h>
 #include <zephyr/logging/log.h>
 LOG_MODULE_REGISTER(HORUSV2, CONFIG_HORUS_LOG_LEVEL);
-#include <assert.h>
 
 /*---------------------------------------------------------------------------*\
 
@@ -19,64 +22,15 @@ LOG_MODULE_REGISTER(HORUSV2, CONFIG_HORUS_LOG_LEVEL);
 
   [ ] code based interleaver
   [ ] test correction of 1,2 & 3 error patterms    
-
-  1/ Unit test on a PC:
-
-     $ gcc horus_l2.c -o horus_l2 -Wall -DHORUS_L2_UNITTEST
-     $ ./horus_l2
-
-     test 0: 22 bytes of payload data BER: 0.00 errors: 0
-     test 0: 22 bytes of payload data BER: 0.01 errors: 0
-     test 0: 22 bytes of payload data BER: 0.05 errors: 0
-     test 0: 22 bytes of payload data BER: 0.10 errors: 7
-     
-     This indicates it's correcting all channel errors for 22 bytes of
-     payload data, at bit error rate (BER) of 0, 0.01, 0.05.  It falls
-     over at a BER of 0.10 which is expected.
-
-  2/ To build with just the tx function, ie for linking with the payload
-  firmware:
-
-    $ gcc horus_l2.c -c -Wall
-    
-  By default the RX side is #ifdef-ed out, leaving the minimal amount
-  of code for tx.
-
-  3/ Generate some tx_bits as input for testing with fsk_horus:
- 
-    $ gcc horus_l2.c -o horus_l2 -Wall -DGEN_TX_BITS -DSCRAMBLER
-    $ ./horus_l2
-    $ more ../octave/horus_tx_bits_binary.txt
-   
-  4/ Unit testing interleaver:
-
-    $ gcc horus_l2.c -o horus_l2 -Wall -DINTERLEAVER -DTEST_INTERLEAVER -DSCRAMBLER
-
-  5/ Compile for use as decoder called by  fsk_horus.m and fsk_horus_stream.m:
-
-    $ gcc horus_l2.c -o horus_l2 -Wall -DDEC_RX_BITS -DHORUS_L2_RX
-
 \*---------------------------------------------------------------------------*/
-
-#include "f_core/protocols/horus/horus.h"
 
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
-#define INTERLEAVER
-#define SCRAMBLER
-
 static char uw[] = {'$', '$'};
 
 /* Function Prototypes ------------------------------------------------*/
-
-int32_t get_syndrome(int32_t pattern);
-
-int golay23_init(void);
-
-int golay23_decode(int received_codeword);
-
 unsigned short gen_crc16(unsigned char *data_p, unsigned char length);
 
 void interleave(unsigned char *inout, int nbytes, int dir);
@@ -188,7 +142,7 @@ int horus_l2_encode_tx_packet(unsigned char *output_tx_data, unsigned char *inpu
                 } else {
                     /* OK we have a full byte ready */
                     *pout = paritybyte;
-                    LOG_INF("      Write paritybyte: 0x%02x", paritybyte);
+                    LOG_INF("\t\tWrite paritybyte: 0x%02x", paritybyte);
                     pout++;
                     paritybyte = 0;
                 }
@@ -203,22 +157,22 @@ int horus_l2_encode_tx_packet(unsigned char *output_tx_data, unsigned char *inpu
     if (ningolay % 12) {
         ingolay >>= 1;
         int32_t golayparity = get_syndrome(ingolay << 12);
-        LOG_INF("  ningolay: %d ingolay: 0x%04x", ningolay, ingolay);
-        LOG_INF("  golayparity: 0x%04x", golayparity);
+        LOG_INF("\tningolay: %d ingolay: 0x%04x", ningolay, ingolay);
+        LOG_INF("\tgolayparity: 0x%04x", golayparity);
 
         /* write parity bits to output data */
 
         for (int i = 0; i < 11; i++) {
             int golayparitybit = (golayparity >> (10 - i)) & 0x1;
             paritybyte = paritybyte | golayparitybit;
-            LOG_DBG("    i: %d golayparitybit: %d paritybyte: 0x%02x", i, golayparitybit, paritybyte);
+            LOG_DBG("\ti: %d golayparitybit: %d paritybyte: 0x%02x", i, golayparitybit, paritybyte);
             nparitybits++;
             if (nparitybits % 8) {
                 paritybyte <<= 1;
             } else {
                 /* OK we have a full byte ready */
                 *pout++ = (unsigned char) paritybyte;
-                LOG_INF("      Write paritybyte: 0x%02x", paritybyte);
+                LOG_INF("\t\tWrite paritybyte: 0x%02x", paritybyte);
                 paritybyte = 0;
             }
         }
@@ -229,7 +183,7 @@ int horus_l2_encode_tx_packet(unsigned char *output_tx_data, unsigned char *inpu
     if (nparitybits % 8) {
         paritybyte <<= 7 - (nparitybits % 8); // use MS bits first
         *pout++ = (unsigned char) paritybyte;
-        LOG_INF("      Write last paritybyte: 0x%02x nparitybits: %d", paritybyte, nparitybits);
+        LOG_INF("\t\tWrite last paritybyte: 0x%02x nparitybits: %d", paritybyte, nparitybits);
     }
 
     // LOG_INF(pout - output_tx_data: %ld num_tx_data_bytes: %d", pout - output_tx_data, num_tx_data_bytes);
@@ -314,17 +268,17 @@ void horus_l2_decode_rx_packet(unsigned char *output_payload_data, unsigned char
                 if ((nparitybits % 8) == 0) {
                     /* OK grab a new byte */
                     paritybyte = *pin++;
-                    LOG_INF("  Read paritybyte: 0x%02x", paritybyte);
+                    LOG_INF("\tRead paritybyte: 0x%02x", paritybyte);
                 }
             }
 
-            LOG_INF("  golay code word: 0x%04x", ingolay);
-            LOG_INF("  golay decode...: 0x%04x", golay23_decode(ingolay));
+            LOG_INF("\tgolay code word: 0x%04x", ingolay);
+            LOG_INF("\tgolay decode...: 0x%04x", golay23_decode(ingolay));
 
             /* write decoded/error corrected bits to output payload data */
 
             int outdata = golay23_decode(ingolay) >> 11;
-            LOG_INF("  outdata...: 0x%04x", outdata);
+            LOG_INF("\toutdata...: 0x%04x", outdata);
 
             for (int i = 0; i < 12; i++) {
                 shift = 11 - i;
@@ -342,7 +296,7 @@ void horus_l2_decode_rx_packet(unsigned char *output_payload_data, unsigned char
 
             ingolay = 0;
         }
-    } /* while(.... */
+    }
 
     LOG_INF("finishing up .....");
 
@@ -359,7 +313,7 @@ void horus_l2_decode_rx_packet(unsigned char *output_payload_data, unsigned char
             if ((nparitybits % 8) == 0) {
                 /* OK grab a new byte */
                 paritybyte = *pin++;
-                LOG_INF("  Read paritybyte: 0x%02x", paritybyte);
+                LOG_INF("\tRead paritybyte: 0x%02x", paritybyte);
             }
         }
 
@@ -384,7 +338,7 @@ void horus_l2_decode_rx_packet(unsigned char *output_payload_data, unsigned char
             if (noutbits % 8) {
                 outbyte <<= 1;
             } else {
-                LOG_INF("  output payload byte: 0x%02x", outbyte);
+                LOG_INF("\toutput payload byte: 0x%02x", outbyte);
                 *pout++ = outbyte;
                 outbyte = 0;
             }
@@ -407,28 +361,24 @@ uint16_t primes[] = {2,   3,   5,   7,   11,  13,  17,  19,  23,  29,  31,  37, 
 
 void interleave(unsigned char *inout, int nbytes, int dir) {
     uint16_t nbits = (uint16_t) nbytes * 8;
-    uint32_t i, j, n, ibit, ibyte, ishift, jbyte, jshift;
-    uint32_t b;
-    unsigned char out[nbytes];
-
-    memset(out, 0, nbytes);
+    unsigned char out[nbytes] = {};
 
     /* b chosen to be co-prime with nbits, I'm cheating by just finding the
        nearest prime to nbits.  It also uses storage, is run on every call,
        and has an upper limit.  Oh Well, still seems to interleave OK. */
-    i = 1;
+    uint32_t i = 1;
     uint16_t imax = sizeof(primes) / sizeof(uint16_t);
     while ((primes[i] < nbits) && (i < imax)) i++;
-    b = primes[i - 1];
+    uint32_t b = primes[i - 1];
 
-    for (n = 0; n < nbits; n++) {
+    for (uint32_t n = 0; n < nbits; n++) {
 
         /*
           "On the Analysis and Design of Good Algebraic Interleavers", Xie et al,eq (5)
         */
 
-        i = n;
-        j = (b * i) % nbits;
+        uint32_t i = n;
+        uint32_t j = (b * i) % nbits;
 
         if (dir) {
             uint16_t tmp = j;
@@ -440,12 +390,12 @@ void interleave(unsigned char *inout, int nbytes, int dir) {
 
         /* read bit i and write to bit j postion */
 
-        ibyte = i / 8;
-        ishift = i % 8;
-        ibit = (inout[ibyte] >> ishift) & 0x1;
+        uint32_t ibyte = i / 8;
+        uint32_t ishift = i % 8;
+        uint32_t ibit = (inout[ibyte] >> ishift) & 0x1;
 
-        jbyte = j / 8;
-        jshift = j % 8;
+        uint32_t jbyte = j / 8;
+        uint32_t jshift = j % 8;
 
         /* write jbit to ibit position */
 
@@ -501,273 +451,7 @@ void scramble(unsigned char *inout, int nbytes) {
 
 #endif
 
-/*---------------------------------------------------------------------------*\
-
-                                   GOLAY FUNCTIONS
-
-\*---------------------------------------------------------------------------*/
-
-/* File:    golay23.c
- * Title:   Encoder/decoder for a binary (23,12,7) Golay code
- * Author:  Robert Morelos-Zaragoza (robert@spectra.eng.hawaii.edu)
- * Date:    August 1994
- *
- * The binary (23,12,7) Golay code is an example of a perfect code, that is,
- * the number of syndromes equals the number of correctable error patterns.
- * The minimum distance is 7, so all error patterns of Hamming weight up to
- * 3 can be corrected. The total number of these error patterns is:
- *
- *       Number of errors         Number of patterns
- *       ----------------         ------------------
- *              0                         1
- *              1                        23
- *              2                       253
- *              3                      1771
- *                                     ----
- *    Total number of error patterns = 2048 = 2^{11} = number of syndromes
- *                                               --
- *                number of redundant bits -------^
- *
- * Because of its relatively low length (23), dimension (12) and number of
- * redundant bits (11), the binary (23,12,7) Golay code can be encoded and
- * decoded simply by using look-up tables. The program below uses a 16K
- * encoding table and an 8K decoding table.
- *
- * For more information, suggestions, or other ideas on implementing error
- * correcting codes, please contact me at (I'm temporarily in Japan, but
- * below is my U.S. address):
- *
- *                    Robert Morelos-Zaragoza
- *                    770 S. Post Oak Ln. #200
- *                      Houston, Texas 77056
- *
- *             email: robert@spectra.eng.hawaii.edu
- *
- *       Homework: Add an overall parity-check bit to get the (24,12,8)
- *                 extended Golay code.
- *
- * COPYRIGHT NOTICE: This computer program is free for non-commercial purposes.
- * You may implement this program for any non-commercial application. You may
- * also implement this program for commercial purposes, provided that you
- * obtain my written permission. Any modification of this program is covered
- * by this copyright.
- *
- * ==   Copyright (c) 1994  Robert Morelos-Zaragoza. All rights reserved.   ==
- */
-
-#define X22    0x00400000 /* vector representation of X^{22} */
-#define X11    0x00000800 /* vector representation of X^{11} */
-#define MASK12 0xfffff800 /* auxiliary vector for testing */
-#define GENPOL 0x00000c75 /* generator polinomial, g(x) */
-
-/* Global variables:
- *
- * pattern = error pattern, or information, or received vector
- * encoding_table[] = encoding table
- * decoding_table[] = decoding table
- * data = information bits, i(x)
- * codeword = code bits = x^{11}i(x) + (x^{11}i(x) mod g(x))
- * numerr = number of errors = Hamming weight of error polynomial e(x)
- * position[] = error positions in the vector representation of e(x)
- * recd = representation of corrupted received polynomial r(x) = c(x) + e(x)
- * decerror = number of decoding errors
- * a[] = auxiliary array to generate correctable error patterns
- */
-
-static int inited = 0;
-
-static int encoding_table[4096], decoding_table[2048];
-
-static int arr2int(int a[], int r)
-/*
- * Convert a binary vector of Hamming weight r, and nonzero positions in
- * array a[1]...a[r], to a long integer \sum_{i=1}^r 2^{a[i]-1}.
- */
-{
-    int i;
-    long mul, result = 0, temp;
-
-    for (i = 1; i <= r; i++) {
-        mul = 1;
-        temp = a[i] - 1;
-        while (temp--) mul = mul << 1;
-        result += mul;
-    }
-    return (result);
-}
-
-void nextcomb(int n, int r, int a[])
-/*
- * Calculate next r-combination of an n-set.
- */
-{
-    int i, j;
-
-    a[r]++;
-    if (a[r] <= n) return;
-    j = r - 1;
-    while (a[j] == n - r + j) j--;
-    for (i = r; i >= j; i--) a[i] = a[j] + i - j + 1;
-    return;
-}
-
-int32_t get_syndrome(int32_t pattern)
-/*
- * Compute the syndrome corresponding to the given pattern, i.e., the
- * remainder after dividing the pattern (when considering it as the vector
- * representation of a polynomial) by the generator polynomial, GENPOL.
- * In the program this pattern has several meanings: (1) pattern = infomation
- * bits, when constructing the encoding table; (2) pattern = error pattern,
- * when constructing the decoding table; and (3) pattern = received vector, to
- * obtain its syndrome in decoding.
- */
-{
-    int32_t aux = X22;
-
-    if (pattern >= X11)
-        while (pattern & MASK12) {
-            while (!(aux & pattern)) aux = aux >> 1;
-            pattern ^= (aux / X11) * GENPOL;
-        }
-    return (pattern);
-}
-
-/*---------------------------------------------------------------------------*\
-
-  FUNCTION....: golay23_init()
-  AUTHOR......: David Rowe
-  DATE CREATED: 3 March 2013
-
-  Call this once when you start your program to init the Golay tables.
-
-\*---------------------------------------------------------------------------*/
-
-int golay23_init(void) {
-    int i;
-    long temp;
-    int a[4];
-    int pattern;
-
-    /*
-    * ---------------------------------------------------------------------
-    *                  Generate ENCODING TABLE
-    *
-    * An entry to the table is an information vector, a 32-bit integer,
-    * whose 12 least significant positions are the information bits. The
-    * resulting value is a codeword in the (23,12,7) Golay code: A 32-bit
-    * integer whose 23 least significant bits are coded bits: Of these, the
-    * 12 most significant bits are information bits and the 11 least
-    * significant bits are redundant bits (systematic encoding).
-    * ---------------------------------------------------------------------
-    */
-    for (pattern = 0; pattern < 4096; pattern++) {
-        temp = pattern << 11;                                /* multiply information by X^{11} */
-        encoding_table[pattern] = temp + get_syndrome(temp); /* add redundancy */
-    }
-
-    /*
-    * ---------------------------------------------------------------------
-    *                  Generate DECODING TABLE
-    *
-    * An entry to the decoding table is a syndrome and the resulting value
-    * is the most likely error pattern. First an error pattern is generated.
-    * Then its syndrome is calculated and used as a pointer to the table
-    * where the error pattern value is stored.
-    * ---------------------------------------------------------------------
-    *
-    * (1) Error patterns of WEIGHT 1 (SINGLE ERRORS)
-    */
-    decoding_table[0] = 0;
-    decoding_table[1] = 1;
-    temp = 1;
-    for (i = 2; i <= 23; i++) {
-        temp *= 2;
-        decoding_table[get_syndrome(temp)] = temp;
-    }
-    /*
-    * (2) Error patterns of WEIGHT 2 (DOUBLE ERRORS)
-    */
-    a[1] = 1;
-    a[2] = 2;
-    temp = arr2int(a, 2);
-    decoding_table[get_syndrome(temp)] = temp;
-    for (i = 1; i < 253; i++) {
-        nextcomb(23, 2, a);
-        temp = arr2int(a, 2);
-        decoding_table[get_syndrome(temp)] = temp;
-    }
-    /*
-    * (3) Error patterns of WEIGHT 3 (TRIPLE ERRORS)
-    */
-    a[1] = 1;
-    a[2] = 2;
-    a[3] = 3;
-    temp = arr2int(a, 3);
-    decoding_table[get_syndrome(temp)] = temp;
-    for (i = 1; i < 1771; i++) {
-        nextcomb(23, 3, a);
-        temp = arr2int(a, 3);
-        decoding_table[get_syndrome(temp)] = temp;
-    }
-    inited = 1;
-    return 0;
-}
-
-/*---------------------------------------------------------------------------*\
-
-  FUNCTION....: golay23_encode()
-  AUTHOR......: David Rowe
-  DATE CREATED: 3 March 2013
-
-  Given 12 bits of data retiurns a 23 bit codeword for transmission
-  over the channel.
-
-\*---------------------------------------------------------------------------*/
-
-int golay23_encode(int data) {
-    assert(inited);
-    assert(data <= 0xfff);
-
-    return encoding_table[data];
-}
-
-#ifdef CONFIG_HORUSV2_RX
-
-/*---------------------------------------------------------------------------*\
-
-  FUNCTION....: golay23_decode()
-  AUTHOR......: David Rowe
-  DATE CREATED: 3 March 2013
-
-  Given a 23 bit received codeword, returns the 12 bit corrected data.
-
-\*---------------------------------------------------------------------------*/
-
-int golay23_decode(int received_codeword) {
-    assert(inited);
-    assert((received_codeword < (1 << 23)) && (received_codeword >= 0));
-
-    LOG_INF("syndrome: 0x%x", get_syndrome(received_codeword));
-    return received_codeword ^= decoding_table[get_syndrome(received_codeword)];
-}
-
-int golay23_count_errors(int recd_codeword, int corrected_codeword) {
-    int errors = 0;
-    int diff, i;
-
-    diff = recd_codeword ^ corrected_codeword;
-    for (i = 0; i < 23; i++) {
-        if (diff & 0x1) errors++;
-        diff >>= 1;
-    }
-
-    return errors;
-}
-
-#endif
-
 // from http://stackoverflow.com/questions/10564491/function-to-calculate-a-crc16-checksum
-
 unsigned short gen_crc16(unsigned char *data_p, unsigned char length) {
     unsigned char x;
     unsigned short crc = 0xFFFF;
@@ -802,9 +486,6 @@ bool horusv2_checksum_verify(const struct horus_packet_v2 *input_packet) {
     return checksum == input_packet->checksum;
 }
 #endif
-
-// Initialize the golay tables
-SYS_INIT(golay23_init, POST_KERNEL, 0);
 
 #ifdef __cplusplus
 }
